# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Giveaway Tool application. All endpoints require Telegram Mini App init_data for authentication.",
        "title": "Giveaway Tool API",
        "contact": {
            "name": "API Support",
            "url": "https://t.me/seinarukiro",
            "email": "seinarukiro@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/giveaways": {
            "get": {
                "security": [
                    {
                        "TelegramInitData": []
                    }
                ],
                "description": "Get list of active and completed giveaways created by the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "giveaways"
                ],
                "summary": "Get my giveaways",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GiveawayResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TelegramInitData": []
                    }
                ],
                "description": "Create a new giveaway with specified parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "giveaways"
                ],
                "summary": "Create new giveaway",
                "parameters": [
                    {
                        "description": "Giveaway creation data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GiveawayCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GiveawayResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data (duration must be at least 5 seconds in debug mode or 5 minutes in release mode)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - user is banned",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/giveaways/my/active": {
            "get": {
                "security": [
                    {
                        "TelegramInitData": []
                    }
                ],
                "description": "Get list of active giveaways created by the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "giveaways"
                ],
                "summary": "Get my active giveaways",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GiveawayDetailedResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/giveaways/my/history": {
            "get": {
                "security": [
                    {
                        "TelegramInitData": []
                    }
                ],
                "description": "Get list of completed giveaways created by the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "giveaways"
                ],
                "summary": "Get my giveaways history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GiveawayDetailedResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/giveaways/participated/history": {
            "get": {
                "security": [
                    {
                        "TelegramInitData": []
                    }
                ],
                "description": "Get list of completed giveaways where the current user participated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "giveaways"
                ],
                "summary": "Get participation history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GiveawayDetailedResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/giveaways/prizes/custom": {
            "post": {
                "security": [
                    {
                        "TelegramInitData": []
                    }
                ],
                "description": "Create a new custom prize",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "giveaways"
                ],
                "summary": "Create custom prize",
                "parameters": [
                    {
                        "description": "Custom prize data",
                        "name": "prize",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CustomPrizeCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Prize"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/giveaways/prizes/templates": {
            "get": {
                "security": [
                    {
                        "TelegramInitData": []
                    }
                ],
                "description": "Get list of available prize templates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "giveaways"
                ],
                "summary": "Get prize templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PrizeTemplate"
                            }
                        }
                    }
                }
            }
        },
...
        "/giveaways/{id}/join": {
            "post": {
                "security": [
                    {
                        "TelegramInitData": []
                    }
                ],
                "description": "Join a giveaway as a participant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "giveaways"
                ],
                "summary": "Join giveaway",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Giveaway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully joined"
                    },
                    "400": {
                        "description": "Giveaway ended or max participants reached",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CustomPrizeCreate": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 10
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TelegramInitData": {
            "description": "Telegram Mini App init_data for authentication. This header should contain the raw init_data string from Telegram Mini App. Example: query_id=AAHdF6IQAAAAAN0XohDhrOrc\u0026user=%7B%22id%22%3A123456789%2C%22first_name%22%3A%22Test%22%2C%22last_name%22%3A%22User%22%2C%22username%22%3A%22testuser%22%2C%22language_code%22%3A%22en%22%7D\u0026auth_date=1684062091\u0026hash=c5b7e2f670fdf15479c8f0cd2028d88ad449e1c6273b8c7a69bb32f0f6c6",
            "type": "apiKey",
            "name": "init_data",
            "in": "header"
        }
    },
...
}